requirements.txt
Flask
gunicorn
stripe
Mongo